Fix few errors and add project archiving

+++ data_deletion/expired_data.py
@@ -1,5 +1,6 @@
 import glob
 import os
+import uuid
 from datetime import datetime
 from os import listdir
 from os.path import join, isdir
@@ -179,21 +180,31 @@ class DeliveredDataDeleter(Deleter):
         self.raw_data_dir = cfg['data_deletion']['fastqs']
         self.raw_archive_dir = cfg['data_deletion']['fastq_archives']
         self.processed_data_dir = cfg['data_deletion']['processed_data']
+        self.processed_archive_dir = cfg['data_deletion']['processed_archives']
         self.delivered_data_dir = cfg['data_deletion']['delivered_data']
         self.list_samples = []
         if manual_delete is not None:
             self.list_samples = manual_delete
         self.limit_samples = sample_ids
 
+    @staticmethod
+    def _get_sample_from_list(list_samples):
+        max_query = 20
+        manual_samples = []
+        for start in range(0, len(list_samples), max_query):
+
+            manual_samples.extend(rest_communication.get_documents(
+                'aggregate/samples',
+                quiet=True,
+                match={'$or': [{'sample_id': s} for s in list_samples[start:start+max_query]]},
+                paginate=False
+            ))
+        return manual_samples
+
     def deletable_samples(self):
         manual_samples = []
         if self.list_samples:
-            manual_samples = rest_communication.get_documents(
-                'aggregate/samples',
-                quiet=True,
-                match={'$or': [{'sample_id': s} for s in self.list_samples]},
-                paginate=False
-            )
+            manual_samples = self._get_sample_from_list(self.list_samples)
             manual_samples = [ProcessedSample(s) for s in manual_samples]
 
         return sorted(manual_samples + self._auto_deletable_samples(), key=lambda e: e.sample_data['sample_id'])
@@ -216,15 +227,21 @@ class DeliveredDataDeleter(Deleter):
                 samples.append(s)
         return samples
 
+    def _move_to_unique_file_name(self, source, dest_dir):
+        source_name = os.path.basename(source)
+        dest = join(dest_dir, str(uuid.uuid4()) + '_' + source_name)
+        self._execute('mv %s %s'%(source, dest))
+
     def setup_samples_for_deletion(self, samples, dry_run):
         total_size_to_delete = 0
         for s in samples:
             total_size_to_delete += s.size_of_files
-            deletable_data = join(self.deletion_dir, s.sample_id)
+            deletable_data_dir = join(self.deletion_dir, s.sample_id)
             if not dry_run:
-                s.list_of_files
-                self._execute('mkdir -p ' + deletable_data)
-                self._execute('mv %s %s' % (' '.join(s.list_of_files), deletable_data))
+                if len(s.list_of_files):
+                    self._execute('mkdir -p ' + deletable_data_dir)
+                    for f in s.list_of_files:
+                        self._move_to_unique_file_name(f, deletable_data_dir)
             else:
                 self.info(
                     'Sample %s has %s files to delete (%.2f G)\n%s',
@@ -233,7 +250,8 @@ class DeliveredDataDeleter(Deleter):
                     s.size_of_files/1000000000,
                     '\n'.join(s.list_of_files)
                 )
-                self.info('Will run: mv %s %s' % (' '.join(s.list_of_files), deletable_data))
+                if len(s.list_of_files):
+                    self.info('Will run: mv %s %s' % (' '.join(s.list_of_files), deletable_data_dir))
         self.info('Will delete %.2f G of data' % (total_size_to_delete / 1000000000))
 
 
@@ -254,13 +272,27 @@ class DeliveredDataDeleter(Deleter):
         if all_fastqs:
             return
         # There are no fastqs in that run
-        self.info('Archive '+ run_name)
+        self.info('Archive run '+ run_name)
         # Find the undetermined
         undetermined_fastqs = glob.glob(join(self.raw_data_dir, run_name, 'Undetermined_*.fastq.gz'))
         for f in undetermined_fastqs:
             self._execute('rm '+ f)
-        self._execute('mv %s %s'%(join(self.raw_data_dir, run_name), self.raw_archive_dir))
+        if os.path.exists(join(self.raw_data_dir, run_name)):
+            self._execute('mv %s %s'%(join(self.raw_data_dir, run_name), self.raw_archive_dir))
 
+    def _try_archive_project(self, project_id):
+        sample_records = rest_communication.get_documents(
+            'aggregate/samples',
+            quiet=True,
+            match={'project_id': project_id},
+            paginate=False
+        )
+        status = list(set([s.status for s in sample_records]))
+        if len(status) != 1 or status[0] != 'deleted':
+            return
+        self.info('Archive project '+ project_id)
+        if os.path.exists(join(self.processed_data_dir, project_id)):
+            self._execute('mv %s %s' % (join(self.processed_data_dir, project_id), self.processed_archive_dir))
 
 
     def delete_data(self):
@@ -278,13 +310,19 @@ class DeliveredDataDeleter(Deleter):
         for s in deletable_samples:
             s.mark_as_deleted()
 
-        self.delete_dir(self.deletion_dir)
-        #data has been deleted now clean up empty directories
+        if self.deletion_dir and os.path.isdir(self.deletion_dir):
+            self.delete_dir(self.deletion_dir)
+        # Data has been deleted now clean up empty directories
 
-        #Remove release batch directories if they are empty
-        for folder in set([os.path.dirname(s.released_data_folder) for s in deletable_samples]):
+        # Remove release batch directories if they are empty
+        for folder in set([os.path.dirname(s.released_data_folder) for s in deletable_samples if s.released_data_folder]):
             self._try_delete_empty_dir(folder)
 
-        #Archive run folders if they do not contain any fastq file anymore
+        # Archive run folders if they do not contain any fastq file anymore
         for run_name in set([r[ELEMENT_RUN_NAME] for r in s.run_elements for s in deletable_samples]):
             self._try_archive_run(run_name)
+
+        # Archive project folders if their samples have been deleted
+        for project_id in set([r[ELEMENT_PROJECT_ID] for r in s.run_elements for s in deletable_samples]):
+            self._try_archive_project(project_id)
+
